[style] svg 파일 SvgIcons에 합치기
