--- src/styles/icons/ActivedContent.tsx
@@ -1,10 +1,0 @@
-import { SVGProps } from 'react';
-
-export const ActivedContent = (props: SVGProps<SVGSVGElement>) => (
-  <svg width="25" height="24" viewBox="0 0 25 24" fill="none" xmlns="http://www.w3.org/2000/svg">
-    <path
-      d="M20.2401 5.67007C19.8906 5.24332 19.2538 5.24619 18.8638 5.63622C18.4747 6.0253 18.4793 6.65247 18.8164 7.08735C19.8682 8.44417 20.5 10.1484 20.5 12C20.5 16.42 16.92 20 12.5 20C8.08 20 4.5 16.42 4.5 12C4.5 7.92 7.55 3.99 11.5 3.5V6.09C8.66 6.57 6.5 9.03 6.5 12C6.5 15.31 9.19 18 12.5 18C15.81 18 18.5 15.31 18.5 12C18.5 10.6997 18.0889 9.49762 17.3869 8.51864C17.0652 8.07005 16.4245 8.07548 16.0342 8.46581C15.6445 8.8555 15.6557 9.48248 15.9372 9.9563C16.2938 10.5566 16.5 11.2554 16.5 12C16.5 14.21 14.71 16 12.5 16C10.29 16 8.5 14.21 8.5 12C8.5 10.14 9.78 8.59 11.5 8.14V10.28C10.9 10.63 10.5 11.26 10.5 12C10.5 13.1 11.4 14 12.5 14C13.6 14 14.5 13.1 14.5 12C14.5 11.26 14.1 10.62 13.5 10.28V3C13.5 2.44771 13.0523 2 12.5 2C6.98 2 2.5 6.48 2.5 12C2.5 17.52 6.98 22 12.5 22C18.02 22 22.5 17.52 22.5 12C22.5 9.59886 21.6523 7.3945 20.2401 5.67007Z"
-      fill="#EA464A"
-    />
-  </svg>
-);
