--- src/styles/icons/BackArrowButton.tsx
@@ -1,10 +1,0 @@
-import { SVGProps } from 'react';
-
-export const BackArrowButton = (props: SVGProps<SVGSVGElement>) => (
-  <svg width="30" height="30" viewBox="0 0 30 30" fill="none" xmlns="http://www.w3.org/2000/svg">
-    <path
-      d="M19.5331 9.25919C20.0409 8.82389 20.0398 8.03799 19.5308 7.60411C19.1229 7.25631 18.5224 7.25729 18.1155 7.60644L10.3847 14.2407C9.91957 14.6399 9.91975 15.3599 10.3851 15.7588L18.1249 22.3941C18.5315 22.7427 19.1317 22.7427 19.5384 22.3941C20.0439 21.9607 20.0439 21.1787 19.5384 20.7453L12.8368 15L19.5331 9.25919Z"
-      fill="#757575"
-    />
-  </svg>
-);
