--- src/styles/icons/InactivedFeed.tsx
@@ -1,10 +1,0 @@
-import { SVGProps } from 'react';
-
-export const InactivedFeed = (props: SVGProps<SVGSVGElement>) => (
-  <svg width="25" height="24" viewBox="0 0 25 24" fill="none" xmlns="http://www.w3.org/2000/svg">
-    <path
-      d="M4.5 7C4.5 7.55228 4.94772 8 5.5 8H7.5C8.05228 8 8.5 7.55228 8.5 7V5C8.5 4.44772 8.05228 4 7.5 4H5.5C4.94772 4 4.5 4.44772 4.5 5V7ZM10.5 19C10.5 19.5523 10.9477 20 11.5 20H13.5C14.0523 20 14.5 19.5523 14.5 19V17C14.5 16.4477 14.0523 16 13.5 16H11.5C10.9477 16 10.5 16.4477 10.5 17V19ZM4.5 19C4.5 19.5523 4.94772 20 5.5 20H7.5C8.05228 20 8.5 19.5523 8.5 19V17C8.5 16.4477 8.05228 16 7.5 16H5.5C4.94772 16 4.5 16.4477 4.5 17V19ZM4.5 13C4.5 13.5523 4.94772 14 5.5 14H7.5C8.05228 14 8.5 13.5523 8.5 13V11C8.5 10.4477 8.05228 10 7.5 10H5.5C4.94772 10 4.5 10.4477 4.5 11V13ZM10.5 13C10.5 13.5523 10.9477 14 11.5 14H13.5C14.0523 14 14.5 13.5523 14.5 13V11C14.5 10.4477 14.0523 10 13.5 10H11.5C10.9477 10 10.5 10.4477 10.5 11V13ZM17.5 4C16.9477 4 16.5 4.44772 16.5 5V7C16.5 7.55228 16.9477 8 17.5 8H19.5C20.0523 8 20.5 7.55228 20.5 7V5C20.5 4.44772 20.0523 4 19.5 4H17.5ZM10.5 7C10.5 7.55228 10.9477 8 11.5 8H13.5C14.0523 8 14.5 7.55228 14.5 7V5C14.5 4.44772 14.0523 4 13.5 4H11.5C10.9477 4 10.5 4.44772 10.5 5V7ZM16.5 13C16.5 13.5523 16.9477 14 17.5 14H19.5C20.0523 14 20.5 13.5523 20.5 13V11C20.5 10.4477 20.0523 10 19.5 10H17.5C16.9477 10 16.5 10.4477 16.5 11V13ZM16.5 19C16.5 19.5523 16.9477 20 17.5 20H19.5C20.0523 20 20.5 19.5523 20.5 19V17C20.5 16.4477 20.0523 16 19.5 16H17.5C16.9477 16 16.5 16.4477 16.5 17V19Z"
-      fill="#B7B7B7"
-    />
-  </svg>
-);
