--- src/styles/icons/SearchBackButton.tsx
@@ -1,10 +1,0 @@
-import { SVGProps } from 'react';
-
-export const SaerchBackButton = (props: SVGProps<SVGSVGElement>) => (
-  <svg width="24" height="25" viewBox="0 0 24 25" fill="none" xmlns="http://www.w3.org/2000/svg">
-    <path
-      d="M20 12.9307C20 12.3784 19.5523 11.9307 19 11.9307H7.83L12.5056 8.5133C12.9855 8.1625 13.0381 7.46574 12.6163 7.04687C12.2658 6.69883 11.7092 6.67102 11.3258 6.98238L6.05176 11.2647C5.02222 12.1007 5.07656 13.6889 6.16083 14.4525L11.3602 18.1144C11.73 18.3748 12.2335 18.3314 12.5533 18.0116C12.9856 17.5793 12.8903 16.8547 12.361 16.5488L7.83 13.9307H19C19.5523 13.9307 20 13.483 20 12.9307Z"
-      fill="#757575"
-    />
-  </svg>
-);
