--- src/styles/icons/SearchIcon.tsx
@@ -1,10 +1,0 @@
-import { SVGProps } from 'react';
-
-export const SearchWhiteIcon = (props: SVGProps<SVGSVGElement>) => (
-  <svg width="26" height="26" viewBox="0 0 26 26" fill="none" xmlns="http://www.w3.org/2000/svg">
-    <path
-      d="M17.2904 15.3819H16.4559L16.1601 15.0967C17.1953 13.8925 17.8186 12.3291 17.8186 10.6284C17.8186 6.83615 14.7446 3.76221 10.9524 3.76221C7.16013 3.76221 4.08618 6.83615 4.08618 10.6284C4.08618 14.4207 7.16013 17.4946 10.9524 17.4946C12.6531 17.4946 14.2165 16.8714 15.4207 15.8362L15.7059 16.1319V16.9664L20.9876 22.2376L22.5615 20.6636L17.2904 15.3819ZM10.9524 15.3819C8.3221 15.3819 6.19886 13.2587 6.19886 10.6284C6.19886 7.99812 8.3221 5.87488 10.9524 5.87488C13.5827 5.87488 15.7059 7.99812 15.7059 10.6284C15.7059 13.2587 13.5827 15.3819 10.9524 15.3819Z"
-      fill="white"
-    />
-  </svg>
-);
