--- src/styles/icons/SeeMore.tsx
@@ -1,14 +1,0 @@
-import { SVGProps } from 'react';
-
-export const SeeMore = (props: SVGProps<SVGSVGElement>) => (
-  <svg width="44" height="22" viewBox="0 0 44 22" fill="none" xmlns="http://www.w3.org/2000/svg">
-    <path
-      d="M1.008 12.108H1.908C3.876 12.108 5.16 12.06 6.6 11.796L6.756 13.08C5.268 13.356 3.936 13.416 1.908 13.416H1.008V12.108ZM1.008 5.94H6.036V7.212H2.604V12.6H1.008V5.94ZM8.22 4.956H9.816V16.104H8.22V4.956ZM5.496 8.784H8.892V10.08H5.496V8.784ZM11.1711 13.488H21.2271V14.784H11.1711V13.488ZM15.3831 11.112H16.9791V13.8H15.3831V11.112ZM12.2271 5.664H13.8111V7.368H18.5511V5.664H20.1351V11.448H12.2271V5.664ZM13.8111 8.616V10.176H18.5511V8.616H13.8111ZM29.4941 4.956H31.1021V16.068H29.4941V4.956ZM26.2181 6.108H27.8021C27.8021 9.504 26.6501 12.288 22.7021 14.196L21.8741 12.924C25.0661 11.388 26.2181 9.372 26.2181 6.396V6.108ZM22.4621 6.108H26.9741V7.368H22.4621V6.108Z"
-      fill="#959595"
-    />
-    <path
-      d="M38.5136 14.2414C38.241 14.4824 38.2416 14.9078 38.5148 15.1481C38.7436 15.3492 39.0863 15.3486 39.3143 15.1467L43.1534 11.7492C43.6032 11.3511 43.6031 10.6491 43.153 10.2512L39.3093 6.85296C39.0813 6.65139 38.7388 6.6514 38.5108 6.85296C38.2394 7.09287 38.2394 7.51622 38.5108 7.75613L42.1799 11L38.5136 14.2414Z"
-      fill="#959595"
-    />
-  </svg>
-);
