+++ src/styles/icons/SvgIcons.tsx
@@ -1,5 +1,16 @@
 import { SVGProps } from 'react';
 
+export const MainIcon = (props: SVGProps<SVGSVGElement>) => (
+  <svg width="75" height="81" viewBox="0 0 75 81" fill="none" xmlns="http://www.w3.org/2000/svg">
+    <path
+      fill-rule="evenodd"
+      clip-rule="evenodd"
+      d="M19.3453 70.5981C13.169 67.1614 7.89641 61.9378 4.38909 55.199C-5.14563 36.8793 1.97596 14.2989 20.2956 4.76415C38.6153 -4.77058 61.1957 2.35102 70.7305 20.6707C80.0789 38.6324 73.4155 60.69 55.8849 70.5318L60.9059 80.1791C60.9059 80.1791 54.442 80.8236 44.1673 79.3718C41.7141 79.0252 36.8509 78.1494 32.9503 76.7121C27.4392 74.6813 19.2064 70.6051 19.2064 70.6051L19.3453 70.5981Z"
+      fill="#EA464A"
+    />
+  </svg>
+);
+
 export const BackArrowButton = (props: SVGProps<SVGSVGElement>) => (
   <svg width="26" height="28" viewBox="0 0 26 28" fill="none" xmlns="http://www.w3.org/2000/svg">
     <path
@@ -42,3 +53,116 @@ export const DropDownIcon = (props: SVGProps<SVGSVGElement>) => {
     <path d="M5.25 7.5L9 11.25L12.75 7.5H5.25Z" fill="#1C1B1F" />
   </svg>;
 };
+
+export const SeeMore = (props: SVGProps<SVGSVGElement>) => (
+  <svg width="44" height="22" viewBox="0 0 44 22" fill="none" xmlns="http://www.w3.org/2000/svg">
+    <path
+      d="M1.008 12.108H1.908C3.876 12.108 5.16 12.06 6.6 11.796L6.756 13.08C5.268 13.356 3.936 13.416 1.908 13.416H1.008V12.108ZM1.008 5.94H6.036V7.212H2.604V12.6H1.008V5.94ZM8.22 4.956H9.816V16.104H8.22V4.956ZM5.496 8.784H8.892V10.08H5.496V8.784ZM11.1711 13.488H21.2271V14.784H11.1711V13.488ZM15.3831 11.112H16.9791V13.8H15.3831V11.112ZM12.2271 5.664H13.8111V7.368H18.5511V5.664H20.1351V11.448H12.2271V5.664ZM13.8111 8.616V10.176H18.5511V8.616H13.8111ZM29.4941 4.956H31.1021V16.068H29.4941V4.956ZM26.2181 6.108H27.8021C27.8021 9.504 26.6501 12.288 22.7021 14.196L21.8741 12.924C25.0661 11.388 26.2181 9.372 26.2181 6.396V6.108ZM22.4621 6.108H26.9741V7.368H22.4621V6.108Z"
+      fill="#959595"
+    />
+    <path
+      d="M38.5136 14.2414C38.241 14.4824 38.2416 14.9078 38.5148 15.1481C38.7436 15.3492 39.0863 15.3486 39.3143 15.1467L43.1534 11.7492C43.6032 11.3511 43.6031 10.6491 43.153 10.2512L39.3093 6.85296C39.0813 6.65139 38.7388 6.6514 38.5108 6.85296C38.2394 7.09287 38.2394 7.51622 38.5108 7.75613L42.1799 11L38.5136 14.2414Z"
+      fill="#959595"
+    />
+  </svg>
+);
+
+export const SearchWhiteIcon = (props: SVGProps<SVGSVGElement>) => (
+  <svg width="26" height="26" viewBox="0 0 26 26" fill="none" xmlns="http://www.w3.org/2000/svg">
+    <path
+      d="M17.2904 15.3819H16.4559L16.1601 15.0967C17.1953 13.8925 17.8186 12.3291 17.8186 10.6284C17.8186 6.83615 14.7446 3.76221 10.9524 3.76221C7.16013 3.76221 4.08618 6.83615 4.08618 10.6284C4.08618 14.4207 7.16013 17.4946 10.9524 17.4946C12.6531 17.4946 14.2165 16.8714 15.4207 15.8362L15.7059 16.1319V16.9664L20.9876 22.2376L22.5615 20.6636L17.2904 15.3819ZM10.9524 15.3819C8.3221 15.3819 6.19886 13.2587 6.19886 10.6284C6.19886 7.99812 8.3221 5.87488 10.9524 5.87488C13.5827 5.87488 15.7059 7.99812 15.7059 10.6284C15.7059 13.2587 13.5827 15.3819 10.9524 15.3819Z"
+      fill="white"
+    />
+  </svg>
+);
+
+export const SaerchBackButton = (props: SVGProps<SVGSVGElement>) => (
+  <svg width="24" height="25" viewBox="0 0 24 25" fill="none" xmlns="http://www.w3.org/2000/svg">
+    <path
+      d="M20 12.9307C20 12.3784 19.5523 11.9307 19 11.9307H7.83L12.5056 8.5133C12.9855 8.1625 13.0381 7.46574 12.6163 7.04687C12.2658 6.69883 11.7092 6.67102 11.3258 6.98238L6.05176 11.2647C5.02222 12.1007 5.07656 13.6889 6.16083 14.4525L11.3602 18.1144C11.73 18.3748 12.2335 18.3314 12.5533 18.0116C12.9856 17.5793 12.8903 16.8547 12.361 16.5488L7.83 13.9307H19C19.5523 13.9307 20 13.483 20 12.9307Z"
+      fill="#757575"
+    />
+  </svg>
+);
+
+export const PostIcon = (props: SVGProps<SVGSVGElement>) => (
+  <svg width="53" height="65" viewBox="0 0 53 65" fill="none" xmlns="http://www.w3.org/2000/svg">
+    <g filter="url(#filter0_d_37_363)">
+      <circle cx="26.3667" cy="34.5667" r="22.3667" fill="#EA464A" />
+      <path
+        d="M24.7309 42.9782V23.9157H27.9657V42.9782H24.7309ZM16.817 35.0644V31.8295H35.8795V35.0644H16.817Z"
+        fill="white"
+      />
+    </g>
+    <defs>
+      <filter
+        id="filter0_d_37_363"
+        x="0"
+        y="12.2"
+        width="52.7334"
+        height="52.7333"
+        filterUnits="userSpaceOnUse"
+        colorInterpolationFilters="sRGB"
+      >
+        <feFlood floodOpacity="0" result="BackgroundImageFix" />
+        <feColorMatrix
+          in="SourceAlpha"
+          type="matrix"
+          values="0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 127 0"
+          result="hardAlpha"
+        />
+        <feOffset dy="4" />
+        <feGaussianBlur stdDeviation="2" />
+        <feComposite in2="hardAlpha" operator="out" />
+        <feColorMatrix type="matrix" values="0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0.12 0" />
+        <feBlend mode="normal" in2="BackgroundImageFix" result="effect1_dropShadow_37_363" />
+        <feBlend mode="normal" in="SourceGraphic" in2="effect1_dropShadow_37_363" result="shape" />
+      </filter>
+    </defs>
+  </svg>
+);
+
+export const InactivedFeed = (props: SVGProps<SVGSVGElement>) => (
+  <svg width="25" height="24" viewBox="0 0 25 24" fill="none" xmlns="http://www.w3.org/2000/svg">
+    <path
+      d="M4.5 7C4.5 7.55228 4.94772 8 5.5 8H7.5C8.05228 8 8.5 7.55228 8.5 7V5C8.5 4.44772 8.05228 4 7.5 4H5.5C4.94772 4 4.5 4.44772 4.5 5V7ZM10.5 19C10.5 19.5523 10.9477 20 11.5 20H13.5C14.0523 20 14.5 19.5523 14.5 19V17C14.5 16.4477 14.0523 16 13.5 16H11.5C10.9477 16 10.5 16.4477 10.5 17V19ZM4.5 19C4.5 19.5523 4.94772 20 5.5 20H7.5C8.05228 20 8.5 19.5523 8.5 19V17C8.5 16.4477 8.05228 16 7.5 16H5.5C4.94772 16 4.5 16.4477 4.5 17V19ZM4.5 13C4.5 13.5523 4.94772 14 5.5 14H7.5C8.05228 14 8.5 13.5523 8.5 13V11C8.5 10.4477 8.05228 10 7.5 10H5.5C4.94772 10 4.5 10.4477 4.5 11V13ZM10.5 13C10.5 13.5523 10.9477 14 11.5 14H13.5C14.0523 14 14.5 13.5523 14.5 13V11C14.5 10.4477 14.0523 10 13.5 10H11.5C10.9477 10 10.5 10.4477 10.5 11V13ZM17.5 4C16.9477 4 16.5 4.44772 16.5 5V7C16.5 7.55228 16.9477 8 17.5 8H19.5C20.0523 8 20.5 7.55228 20.5 7V5C20.5 4.44772 20.0523 4 19.5 4H17.5ZM10.5 7C10.5 7.55228 10.9477 8 11.5 8H13.5C14.0523 8 14.5 7.55228 14.5 7V5C14.5 4.44772 14.0523 4 13.5 4H11.5C10.9477 4 10.5 4.44772 10.5 5V7ZM16.5 13C16.5 13.5523 16.9477 14 17.5 14H19.5C20.0523 14 20.5 13.5523 20.5 13V11C20.5 10.4477 20.0523 10 19.5 10H17.5C16.9477 10 16.5 10.4477 16.5 11V13ZM16.5 19C16.5 19.5523 16.9477 20 17.5 20H19.5C20.0523 20 20.5 19.5523 20.5 19V17C20.5 16.4477 20.0523 16 19.5 16H17.5C16.9477 16 16.5 16.4477 16.5 17V19Z"
+      fill="#B7B7B7"
+    />
+  </svg>
+);
+
+export const InactivedContent = (props: SVGProps<SVGSVGElement>) => (
+  <svg width="25" height="24" viewBox="0 0 25 24" fill="none" xmlns="http://www.w3.org/2000/svg">
+    <path
+      d="M20.2401 5.67007C19.8906 5.24332 19.2538 5.24619 18.8638 5.63622C18.4747 6.0253 18.4793 6.65247 18.8164 7.08735C19.8682 8.44417 20.5 10.1484 20.5 12C20.5 16.42 16.92 20 12.5 20C8.08 20 4.5 16.42 4.5 12C4.5 7.92 7.55 3.99 11.5 3.5V6.09C8.66 6.57 6.5 9.03 6.5 12C6.5 15.31 9.19 18 12.5 18C15.81 18 18.5 15.31 18.5 12C18.5 10.6997 18.0889 9.49762 17.3869 8.51864C17.0652 8.07005 16.4245 8.07548 16.0342 8.46581C15.6445 8.8555 15.6557 9.48248 15.9372 9.9563C16.2938 10.5566 16.5 11.2554 16.5 12C16.5 14.21 14.71 16 12.5 16C10.29 16 8.5 14.21 8.5 12C8.5 10.14 9.78 8.59 11.5 8.14V10.28C10.9 10.63 10.5 11.26 10.5 12C10.5 13.1 11.4 14 12.5 14C13.6 14 14.5 13.1 14.5 12C14.5 11.26 14.1 10.62 13.5 10.28V3C13.5 2.44771 13.0523 2 12.5 2C6.98 2 2.5 6.48 2.5 12C2.5 17.52 6.98 22 12.5 22C18.02 22 22.5 17.52 22.5 12C22.5 9.59886 21.6523 7.3945 20.2401 5.67007Z"
+      fill="#B7B7B7"
+    />
+  </svg>
+);
+
+export const BackArrowButton2 = (props: SVGProps<SVGSVGElement>) => (
+  <svg width="30" height="30" viewBox="0 0 30 30" fill="none" xmlns="http://www.w3.org/2000/svg">
+    <path
+      d="M19.5331 9.25919C20.0409 8.82389 20.0398 8.03799 19.5308 7.60411C19.1229 7.25631 18.5224 7.25729 18.1155 7.60644L10.3847 14.2407C9.91957 14.6399 9.91975 15.3599 10.3851 15.7588L18.1249 22.3941C18.5315 22.7427 19.1317 22.7427 19.5384 22.3941C20.0439 21.9607 20.0439 21.1787 19.5384 20.7453L12.8368 15L19.5331 9.25919Z"
+      fill="#757575"
+    />
+  </svg>
+);
+
+export const ActivedFeed = (props: SVGProps<SVGSVGElement>) => (
+  <svg width="25" height="24" viewBox="0 0 25 24" fill="none" xmlns="http://www.w3.org/2000/svg">
+    <path
+      d="M4.5 7C4.5 7.55228 4.94772 8 5.5 8H7.5C8.05228 8 8.5 7.55228 8.5 7V5C8.5 4.44772 8.05228 4 7.5 4H5.5C4.94772 4 4.5 4.44772 4.5 5V7ZM10.5 19C10.5 19.5523 10.9477 20 11.5 20H13.5C14.0523 20 14.5 19.5523 14.5 19V17C14.5 16.4477 14.0523 16 13.5 16H11.5C10.9477 16 10.5 16.4477 10.5 17V19ZM4.5 19C4.5 19.5523 4.94772 20 5.5 20H7.5C8.05228 20 8.5 19.5523 8.5 19V17C8.5 16.4477 8.05228 16 7.5 16H5.5C4.94772 16 4.5 16.4477 4.5 17V19ZM4.5 13C4.5 13.5523 4.94772 14 5.5 14H7.5C8.05228 14 8.5 13.5523 8.5 13V11C8.5 10.4477 8.05228 10 7.5 10H5.5C4.94772 10 4.5 10.4477 4.5 11V13ZM10.5 13C10.5 13.5523 10.9477 14 11.5 14H13.5C14.0523 14 14.5 13.5523 14.5 13V11C14.5 10.4477 14.0523 10 13.5 10H11.5C10.9477 10 10.5 10.4477 10.5 11V13ZM17.5 4C16.9477 4 16.5 4.44772 16.5 5V7C16.5 7.55228 16.9477 8 17.5 8H19.5C20.0523 8 20.5 7.55228 20.5 7V5C20.5 4.44772 20.0523 4 19.5 4H17.5ZM10.5 7C10.5 7.55228 10.9477 8 11.5 8H13.5C14.0523 8 14.5 7.55228 14.5 7V5C14.5 4.44772 14.0523 4 13.5 4H11.5C10.9477 4 10.5 4.44772 10.5 5V7ZM16.5 13C16.5 13.5523 16.9477 14 17.5 14H19.5C20.0523 14 20.5 13.5523 20.5 13V11C20.5 10.4477 20.0523 10 19.5 10H17.5C16.9477 10 16.5 10.4477 16.5 11V13ZM16.5 19C16.5 19.5523 16.9477 20 17.5 20H19.5C20.0523 20 20.5 19.5523 20.5 19V17C20.5 16.4477 20.0523 16 19.5 16H17.5C16.9477 16 16.5 16.4477 16.5 17V19Z"
+      fill="#EA464A"
+    />
+  </svg>
+);
+
+export const ActivedContent = (props: SVGProps<SVGSVGElement>) => (
+  <svg width="25" height="24" viewBox="0 0 25 24" fill="none" xmlns="http://www.w3.org/2000/svg">
+    <path
+      d="M20.2401 5.67007C19.8906 5.24332 19.2538 5.24619 18.8638 5.63622C18.4747 6.0253 18.4793 6.65247 18.8164 7.08735C19.8682 8.44417 20.5 10.1484 20.5 12C20.5 16.42 16.92 20 12.5 20C8.08 20 4.5 16.42 4.5 12C4.5 7.92 7.55 3.99 11.5 3.5V6.09C8.66 6.57 6.5 9.03 6.5 12C6.5 15.31 9.19 18 12.5 18C15.81 18 18.5 15.31 18.5 12C18.5 10.6997 18.0889 9.49762 17.3869 8.51864C17.0652 8.07005 16.4245 8.07548 16.0342 8.46581C15.6445 8.8555 15.6557 9.48248 15.9372 9.9563C16.2938 10.5566 16.5 11.2554 16.5 12C16.5 14.21 14.71 16 12.5 16C10.29 16 8.5 14.21 8.5 12C8.5 10.14 9.78 8.59 11.5 8.14V10.28C10.9 10.63 10.5 11.26 10.5 12C10.5 13.1 11.4 14 12.5 14C13.6 14 14.5 13.1 14.5 12C14.5 11.26 14.1 10.62 13.5 10.28V3C13.5 2.44771 13.0523 2 12.5 2C6.98 2 2.5 6.48 2.5 12C2.5 17.52 6.98 22 12.5 22C18.02 22 22.5 17.52 22.5 12C22.5 9.59886 21.6523 7.3945 20.2401 5.67007Z"
+      fill="#EA464A"
+    />
+  </svg>
+);
